Change services to s6-rc
This is meant to address linuxserver/docker-deluge#105 in a more permanent way,
by switching the services over to use the new init system. Instead of
relying on what s6-overlay calls the "legacy" system, which uses rigid
timeouts and sometimes fails to wait for deluge to exit cleanly, s6-rc
will wait as long as it takes for deluge to exit. This means that the
shutdown timeout is determined by the container manager, instead of the
grace variables outlined at https://github.com/just-containers/s6-overlay/blob/master/README.md#customizing-s6-behaviour

The dependency chain is as follows:
1. `deluged-initial-setup`: waits for `legacy-cont-init` and `init-mods`
  - This is the old `30-config` script.
  - The abc user and group are added in legacy `10-adduser`, so that has to run before the `chown` check.
2. `deluged`: waits for `deluged-initial-setup`
3. `deluge-web` and `deluged-set-folders`: wait for `deluge`
  - `deluged-set-folders` is the old `apiping`

I also made the scripts executable, since that was mentioned in the
s6-overlay docs. Sorry if I'm stepping on the toes of PR #155.